requirements.txt
pydantic~=1.7.2
tensorflow~=2.3.1
numpy~=1.19.4
pandas~=1.1.4
nltk~=3.5
fastapi~=0.61.2
databases~=0.4.0
SQLAlchemy~=1.3.20
python-dotenv~=0.15.0
uvicorn~=0.12.2